seperate tests for each api_workflow_client Mixin
