+++ tests/api_workflow/test_api_workflow_upload_embeddings.py
@@ -1,0 +1,28 @@
+import os
+import tempfile
+
+import numpy as np
+
+import lightly
+from tests.api_workflow.mocked_api_workflow_client import MockedApiWorkflowSetup
+
+
+class TestApiWorkflowUploadEmbeddigns(MockedApiWorkflowSetup):
+    def test_upload_embedding(self, n_data: int = 100):
+        # create fake embeddings
+        folder_path = tempfile.mkdtemp()
+        path_to_embeddings = os.path.join(
+            folder_path,
+            'embeddings.csv'
+        )
+        sample_names = [f'img_{i}.jpg' for i in range(n_data)]
+        labels = [0] * len(sample_names)
+        lightly.utils.save_embeddings(
+            path_to_embeddings,
+            np.random.randn(n_data, 16),
+            labels,
+            sample_names
+        )
+
+        # perform the workflow to upload the embeddings
+        self.api_workflow_client.upload_embeddings(path_to_embeddings_csv=path_to_embeddings, name="embedding_xyz")
